+++ geofileops/fileops.py
@@ -2133,8 +2133,7 @@ def remove(path: Union[str, "os.PathLike[Any]"], missing_ok: bool = False):
     lockfile_path.unlink(missing_ok=True)
 
     # Remove the main file
-    if path.exists():
-        path.unlink(missing_ok=missing_ok)
+    path.unlink(missing_ok=missing_ok)
 
     # For some file types, extra files need to be removed
     for suffix in path_info.suffixes_extrafiles:
@@ -2585,7 +2584,7 @@ def copy_layer(
         raise ValueError(f"src file doesn't exist: {src}")
     # If dest file exists already and no append
     if not append and dst.exists():
-        if force is True:
+        if force:
             remove(dst)
         else:
             logger.info(f"Output file exists already, so stop: {dst}")