+++ CHANGELOG.md
@@ -22,7 +22,13 @@
 - Don't convert to multi-part geometries by default in `copy_layer`,... (#570)
 - Improve performance of erase for very complex input geometries (#585)
 - Add configuration option to only warn on dissolve errors (#561)
+- Add some pre-flight checks when geofileops is imported (#573)
 - Add support for renaming layer with only difference in casing (#593)
+
+### Bugs fixed
+
+- Fix `copy_layer` to a gpkg.zip file (#604)
+- Fix `missing_ok` parameter in `remove` being ~ignored (#605)
 
 ## 0.9.1 (2024-07-18)